Removed rush-eval::evaluator::Evaluator wrapper type in favor of simply exposing the Dispatcher directly.
